Factor mocker into several files.
